:sparkles: support third party command
