+++ commands/app.go
@@ -1,7 +1,6 @@
 package commands
 
 import (
-	"log"
 	"net/url"
 	"os"
 	"os/exec"
@@ -12,27 +11,6 @@ import (
 	"github.com/leancloud/lean-cli/version"
 	"github.com/pkg/browser"
 )
-
-func thirdPartyCommand(c *cli.Context, _cmdName string) {
-	cmdName := "lean-" + _cmdName
-
-	// executeble not found:
-	execPath, err := exec.LookPath(cmdName)
-	if e, ok := err.(*exec.Error); ok {
-		if e.Err == exec.ErrNotFound {
-			cli.ShowAppHelp(c)
-			return
-		}
-	}
-	cmd := exec.Command(execPath, c.Args()[1:]...)
-	cmd.Stderr = os.Stderr
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	err = cmd.Run()
-	if err != nil {
-		log.Fatal(err)
-	}
-}
 
 // Run the command line
 func Run(args []string) {
